Optimize happy path case of constantize
The changes to const_defined since Ruby 2.0 allow it to
solve the most basic use cases for constantize.

As discussed in https://groups.google.com/forum/#!topic/rubyonrails-core/JO7KceC8LjY,
there are a few subtle differences between the two implementations.

Luckily, an additional call to const_defined is all that's necessary to identify these edge cases.

We still need to keep the old logic around for those, but for the majority of cases,
const_get should be enough.

Comparison (on a top-level constant):
           const_get:  4016159.6 i/s
            shortcut:  2010652.1 i/s - 2.00x slower
         constantize:  1127093.1 i/s - 3.56x slower